refactor initial single-page confluence support
This commit refactors the initial implementation for single-page
rendering support for the Sphinx Confluence extension. Instead of
introducing a configuration entry `confluence_single_page` to toggle
single-page building, make a dedicated worker `singleconfluence`. This
pattern mimics the same user pattern when a user invokes an `html`
builder or a `singlehtml` builder. A user wishing to build a single
Confluence page would invoke the following:

    sphinx-build -b singleconfluence . _build/singleconfluence -E -a

Note that this commit is an initial rework and does not address various
limitations with references, assets, etc. with a singleconfluence
builder invoke.
